refactor socketAdapterClient
