+++ noah/src/main/scala/edu/uci/ics/cloudberry/noah/feed/FeedSocketAdapterClient.java
@@ -1,9 +1,14 @@
 package edu.uci.ics.cloudberry.noah.feed;
 
 import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.OutputStream;
 import java.net.Socket;
 
-public abstract class FeedSocketAdapterClient {
+public class FeedSocketAdapterClient {
+    private OutputStream out = null;
+    private int recordCount = 0;
+
     protected String adapterUrl;
     protected int port;
     protected int waitMillSecond;
@@ -12,9 +17,43 @@ public abstract class FeedSocketAdapterClient {
 
     protected Socket socket;
 
-    public abstract void initialize() throws IOException;
+    public FeedSocketAdapterClient(String adapterUrl, int port, int batchSize,
+                                         int waitMillSecPerRecord, int maxCount) {
+        this.adapterUrl = adapterUrl;
+        this.port = port;
+        this.maxCount = maxCount;
+        this.waitMillSecond = waitMillSecPerRecord;
+        this.batchSize = batchSize;
+    }
 
-    public abstract void finalize();
+    public void initialize() throws IOException {
+        socket = new Socket(adapterUrl, port);
+        out = socket.getOutputStream();
+    }
+
+    public void finalize() {
+        try {
+            out.close();
+            socket.close();
+        } catch (IOException e) {
+            System.err.println("Problem in closing socket against host " + adapterUrl + " on the port " + port);
+            e.printStackTrace();
+        }
+    }
+
+    public void ingest(String record) throws IOException{
+        recordCount++;
+        System.out.println("send record: " + recordCount);
+        byte[] b = record.replaceAll("\\s+", " ").getBytes();
+        try {
+            out.write(b);
+            if (waitMillSecond >= 1 && recordCount % batchSize == 0) {
+                Thread.currentThread().sleep(waitMillSecond);
+            }
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+    }
 
 
 }
